backend/requirements.txt
# App
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
coverage==5.5

# DB
psycopg2-binary>=2.8

# Testing
pytest==5.4.1
pytest-django==3.8.0